Determine the 'source' port for a tunnel dynamically.
This allows multiple tunnels to exist on the same machine. However, you can't
read the stream until you know the forwarder's port, and the tunnels are created
out-of-band of creating the tickets.

The ticket will be in a state of ambiguity for a moment, before the tunnel kicks
in and sets the port on the Ticket in the database. The browser will need to
wait a moment, then GET the ticket again to find this information.

A better solution would be to have the tunnel created synchronously with the
Ticket. The ticket handler would then not return to the client until it has a
valid port for the ticket.