Always route source through a tunnel, so we can control stream.
This adds the functionality to pause/resume a stream using another TCP tunnel at
the source, but there must be some tricky stuff with RTMP. If it's not enabled
at the start, the nodes never seem to get it. I wonder if they expect the set-up
messages right after the socket is opened.